+++ src/CDItest.cc
@@ -67,7 +67,7 @@ void *CDItest(void *process)
       int streamID2 = cdoStreamOpenWrite(cdoStreamName(1), cdoFiletype());
       pstreamDefVlist(streamID2, vlistID2);
 
-      int gridsize = vlistGridsizeMax(vlistID1);
+      size_t gridsize = vlistGridsizeMax(vlistID1);
       double *array = (double*) Malloc(gridsize*sizeof(double));
 
       int tsID1 = 0;
