+++ src/Magplot.cc
@@ -1162,7 +1162,7 @@ void *Magplot(void *process)
   
   if ( gridtype != GRID_CURVILINEAR ) gridID = gridToCurvilinear(gridID, 1);
 
-  int gridsize = gridInqSize(gridID);
+  size_t gridsize = gridInqSize(gridID);
   int nlon     = gridInqXsize(gridID);
   int nlat     = gridInqYsize(gridID);
   //int nlev     = zaxisInqSize(zaxisID);
